504. Eliminando el Proyecto y sincronizando el State

+++ frontend/src/components/ButtonEliminarProyecto.tsx
@@ -5,19 +5,20 @@ interface ButtonEliminarProyectoProps {
     id: string
 }
 
-export const ButtonEliminarProyecto = ({id}:ButtonEliminarProyectoProps) => {
-    const { eliminarProyecto}= useProyecto()
-   const handleClink = () => {
-    eliminarProyecto(id)
-   }
-
+export const ButtonEliminarProyecto = ({ id }: ButtonEliminarProyectoProps) => {
+    const { eliminarProyecto } = useProyecto()
+    const handleClick = () => {
+        if (confirm('Desea eliminar este proyecto?')) {
+            eliminarProyecto(id)
+        }
+    }
 
     return (
         <>
             <div className=" p-2 rounded-md bg-red-600 text-white cursor-pointer font-semibold transition-all duration-300 hover:shadow-lg">
-                <button 
-                 onClick={handleClink}
-                 type="button" className="flex gap-2">
+                <button
+                    onClick={handleClick}
+                    type="button" className="flex gap-2">
                     <EliminarIcon />
                     Eliminar
                 </button>
